requirements.txt
astropy==6.1.3
astropy-iers-data==0.2025.1.13.0.34.51
bottleneck==1.4.2
brotli-python==1.0.9
cftime==1.6.4
contourpy==1.3.1
cycler==0.11.0
filelock==3.17.0
fonttools==4.55.3
gmpy2==2.2.1
h5py==3.12.1
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib-base==3.10.0
mpmath==1.3.0
netcdf4==1.7.2
networkx==3.4.2
numexpr==2.10.1
numpy==2.0.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==25.1
py-cpuinfo==9.0.0
pyerfa==2.0.1.5
pyparsing==3.2.0
pytables==3.10.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==78.1.1
shapely==2.0.6
six==1.17.0
sympy==1.13.3
threadpoolctl==3.5.0
typing-extensions==4.12.2
typing_extensions==4.12.2
wheel==0.45.1
xarray==2024.11.0